divide into multiple files, other review tweaks
